:sparkles: Add user data to list groups
