+++ backend/src/zac/contrib/objects/checklists/api/serializers.py
@@ -10,13 +10,13 @@ from rest_framework import serializers
 from zgw_consumers.api_models.base import factory
 from zgw_consumers.drf.serializers import APIModelSerializer
 
+from zac.accounts.api.fields import GroupSlugRelatedField, UserSlugRelatedField
 from zac.accounts.models import User
 from zac.contrib.objects.services import (
     fetch_checklist,
     fetch_checklist_object,
     fetch_checklisttype,
 )
-from zac.core.api.fields import GroupSlugRelatedField, UserSlugRelatedField
 from zac.core.models import MetaObjectTypesConfig
 from zac.core.services import (
     create_object,
