+++ backend/src/zac/core/api/fields.py
@@ -1,32 +1,8 @@
-from typing import Type
-
 from django.utils.translation import gettext_lazy as _
 
-from rest_framework import fields, relations
-from rest_framework.serializers import Serializer
-
-from zac.accounts.api.serializers import GroupSerializer, UserSerializer
+from rest_framework import fields
 
 from .validators import ZaakDocumentsValidator
-
-
-class SerializerSlugRelatedField(relations.SlugRelatedField):
-    response_serializer = None
-
-    def get_response_serializer(self, obj: object) -> Type[Serializer]:
-        assert self.response_serializer
-        return self.response_serializer(obj)
-
-    def to_representation(self, obj):
-        return self.get_response_serializer(obj).data
-
-
-class UserSlugRelatedField(SerializerSlugRelatedField):
-    response_serializer = UserSerializer
-
-
-class GroupSlugRelatedField(SerializerSlugRelatedField):
-    response_serializer = GroupSerializer
 
 
 class SelectDocumentsField(fields.ListField):
